GPT4Tetris Python Source Code
This is the first check-in version 1.0 of fully GPT4 generated source code for a Tetris game clone. All of the code in this file was 100% generated by GPT4 via ChatGPT without any code being written by me. All features added and bugs resolved were done by just asking in plain English for the features and describing the bugs best I could visually or by uploading a screen shot of the game showing the bug so that GPT4 could figure it out and fix the code. After conversing back and forth about ~50 times I told it to clean up the source code and give me a final version which I had to click "Continue" multiple times to get but the contents of this file were 100% returned from GPT4 as a single reply in the end. Even the requirements.txt file uploaded was generated by GPT4 from the same conversation. This proves just how advanced GPT's are getting when you can have it completely generate a game from scratch just by asking it for features and describing the features and letting it break up the work load into units until the very end after you do the copy and paste work to generate the final assembly with everything just working.